+++ src/main/java/peaksoft/giftlistj7g2/model/dto/LoginRequest.java
@@ -1,4 +1,4 @@
-package peaksoft.giftlistj7g2.dto;
+package peaksoft.giftlistj7g2.model.dto;
 
 import lombok.Getter;
 import lombok.Setter;
