+++ src/main/java/peaksoft/giftlistj7g2/model/dto/LoginResponse.java
@@ -1,20 +1,19 @@
-package peaksoft.giftlistj7g2.dto;
+package peaksoft.giftlistj7g2.model.dto;
 
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.FieldDefaults;
+import peaksoft.giftlistj7g2.model.enums.Role;
+
+import java.util.List;
 
 @Getter
 @Setter
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
-public class RegisterResponse {
-    Long id;
-    String name;
-    String lastName;
-    String email;
-    int age;
-    String response;
+public class LoginResponse {
+    String token;
+    List<String> roleName;
 }
