+++ src/main/java/peaksoft/giftlistj7g2/model/dto/RegisterRequest.java
@@ -1,18 +1,18 @@
-package peaksoft.giftlistj7g2.dto;
+package peaksoft.giftlistj7g2.model.dto;
 
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.FieldDefaults;
 
-import java.util.List;
-
 @Getter
 @Setter
-@Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
-public class LoginResponse {
-    String token;
-    List<String> roleName;
+public class RegisterRequest {
+    String name;
+    String lastName;
+    String email;
+    int age;
+    String password;
 }
+
