+++ src/main/java/peaksoft/giftlistj7g2/model/dto/RegisterResponse.java
@@ -1,16 +1,20 @@
-package peaksoft.giftlistj7g2.dto;
+package peaksoft.giftlistj7g2.model.dto;
 
 import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.FieldDefaults;
 
 @Getter
 @Setter
+@Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
-public class UserRequest {
+public class RegisterResponse {
+    Long id;
     String name;
-    String lastname;
+    String lastName;
     String email;
     int age;
+    String response;
 }
