+++ src/main/java/peaksoft/giftlistj7g2/model/dto/UserRequest.java
@@ -1,22 +1,16 @@
-package peaksoft.giftlistj7g2.dto;
+package peaksoft.giftlistj7g2.model.dto;
 
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.FieldDefaults;
 
-import java.time.LocalDate;
-
 @Getter
 @Setter
-@Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
-public class UserResponse {
-    Long id;
+public class UserRequest {
     String name;
     String lastname;
     String email;
     int age;
-    LocalDate createDate;
 }
