+++ src/main/java/peaksoft/giftlistj7g2/model/dto/UserResponse.java
@@ -1,18 +1,22 @@
-package peaksoft.giftlistj7g2.dto;
+package peaksoft.giftlistj7g2.model.dto;
 
 import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.FieldDefaults;
 
+import java.time.LocalDate;
+
 @Getter
 @Setter
+@Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
-public class RegisterRequest {
+public class UserResponse {
+    Long id;
     String name;
-    String lastName;
+    String lastname;
     String email;
     int age;
-    String password;
+    LocalDate createDate;
 }
-
