+++ src/main/java/peaksoft/giftlistj7g2/security/UserDetailsServiceImpl.java
@@ -7,7 +7,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-import peaksoft.giftlistj7g2.repository.AuthRepository;
+import peaksoft.giftlistj7g2.model.repository.AuthRepository;
 
 @Service
 @FieldDefaults(level = AccessLevel.PRIVATE)
